Struct rustc_arena::DroplessArena

pub struct DroplessArena {
    pub(crate) start: Cell<*mut u8>,
    pub(crate) end: Cell<*mut u8>,
    pub(crate) chunks: RefCell<Vec<ArenaChunk>>,
}

An arena that can hold objects of multiple different types that impl Copy and/or satisfy !mem::needs_drop.

Fields
start: Cell<*mut u8>

A pointer to the start of the free space.

end: Cell<*mut u8>

A pointer to the end of free space.

The allocation proceeds downwards from the end of the chunk towards the start. (This is slightly simpler and faster than allocating upwards, see https://fitzgeraldnick.com/2019/11/01/always-bump-downwards.html.) When this pointer crosses the start pointer, a new chunk is allocated.

This is kept aligned to DROPLESS_ALIGNMENT.

chunks: RefCell<Vec<ArenaChunk>>

A vector of arena chunks.

Implementations

impl DroplessArena

pub(crate) fn grow(&self, layout: Layout)

pub fn alloc_raw(&self, layout: Layout) -> *mut u8

pub fn alloc<T>(&self, object: T) -> &mut T

pub fn alloc_slice<T>(&self, slice: &[T]) -> &mut [T]
where T: Copy,

Allocates a slice of objects that are copied into the DroplessArena, returning a mutable reference to it. Will panic if passed a zero-sized type.

Panics:

Zero-sized types

Zero-length slices

pub fn contains_slice<T>(&self, slice: &[T]) -> bool

Used by Lift to check whether this slice is allocated in this arena.

pub fn alloc_str(&self, string: &str) -> &str

Allocates a string slice that is copied into the DroplessArena, returning a reference to it. Will panic if passed an empty string.

Panics:

Zero-length string

pub(crate) unsafe fn write_from_iter<T, I: Iterator<Item = T>>( &self, iter: I, len: usize, mem: *mut T, ) -> &mut [T]

Safety

The caller must ensure that mem is valid for writes up to size_of::<T>() * len, and that that memory stays allocated and not shared for the lifetime of self. This must hold even if iter.next() allocates onto self.

pub fn alloc_from_iter<T, I: IntoIterator<Item = T>>(&self, iter: I) -> &mut [T]