Struct rustc_arena::TypedArena

pub struct TypedArena<T> {
    pub(crate) ptr: Cell<*mut T>,
    pub(crate) end: Cell<*mut T>,
    pub(crate) chunks: RefCell<Vec<ArenaChunk<T>>>,
    pub(crate) _own: PhantomData<T>,
}

An arena that can hold objects of only one type.

Fields
ptr: Cell<*mut T>

A pointer to the next object to be allocated.

end: Cell<*mut T>

A pointer to the end of the allocated area. When this pointer is reached, a new chunk is allocated.

chunks: RefCell<Vec<ArenaChunk<T>>>

A vector of arena chunks.

_own: PhantomData<T>

Marker indicating that dropping the arena causes its owned instances of T to be dropped.

Implementations

impl<T> TypedArena<T>

pub fn alloc(&self, object: T) -> &mut T

Allocates an object in the TypedArena, returning a reference to it.

pub(crate) fn can_allocate(&self, additional: usize) -> bool

pub(crate) fn ensure_capacity(&self, additional: usize)

Ensures there’s enough space in the current chunk to fit len objects.

pub(crate) unsafe fn alloc_raw_slice(&self, len: usize) -> *mut T

pub fn alloc_from_iter<I: IntoIterator<Item = T>>(&self, iter: I) -> &mut [T]

pub(crate) fn grow(&self, additional: usize)

Grows the arena.

pub(crate) fn clear_last_chunk(&self, last_chunk: &mut ArenaChunk<T>)